pyproject.toml
[tool.poetry]
name = "kida-backend"
version = "1"
description = "Kids Diary, KiDa"
authors = ["0417taehyun <0417taehyun@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
fastapi = "^0.79.1"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
requests = "^2.28.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
PyMySQL = "^1.0.2"
boto3 = "^1.24.71"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
black = "^22.8.0"
isort = "^5.10.1"
mypy = "^0.971"
flake8 = "^5.0.4"
cryptography = "^38.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
